General XS cleanup
The following changes where applied to the XS code:

- Use SvPVbyte instead of SvPV: it is the only sane way to handle data
passing through the library at the moment. Until now, the wrapper was
not differentiating between perl values stored internally as latin1 or
as utf8, so the same string at the perl level, could result in two
different strings when passed to the library. This change can expose
several bugs in user code.

- Properly coerze write buffers passed from the outside: lvalue
arguments passed as arguments need to be sanitized before they can be
used for output.

- Use typemaps to generate proper and consistent return values.

- Save LIBSSH2_ERROR_EAGAIN errors.

- Handle 64bit numbers correctly: as native IVs in 64bits platforms,
or as NVs on 32bits platforms that is still not completely correct,
but enough for practical pourposes.