General XS cleanup
The following changes where applied to the XS code:

- Use SvPVbyte instead of SvPV: it is the only sane way to handle data
passing through the library at the moment. Until now, the wrapper was
not differentiating between perl values stored internally as latin1 or
as utf8, so the same string at the perl level, could result in two
different strings when passed to the library. This change can expose
several bugs in user code.

- Properly coerze write buffers passed from the outside: lvalue
arguments passed as arguments need to be sanitized before they can be
used for output.

- Use typemaps to generate proper and consistent return values.

- Save LIBSSH2_ERROR_EAGAIN errors.

- Handle 64bit numbers correctly: as native IVs in 64bits platforms,
or as NVs on 32bits platforms that is still not completely correct,
but enough for practical pourposes.

- Break setenv in two, a simpler part in XS that just sets one
environment variable, and another part in Perl that can handle several
ones. Now, instead of returning the number of variables sucessfully
set, the method just returns true or undef to indicate that all the
variables were set correctly or that at least one failed respectively.

- Break session method "method" in two parts, one in Perl able to
accept a list and another simpler one in the XS side that only accepts
a scalar.

- KnownHost handling methods were dying on error, now they just return
undef for consistency with the other parts of the module.